+++ src/Pages/CategoryLeaderboardPage/Components/RankRow.js
@@ -3,9 +3,9 @@ import { TouchableOpacity, View, useColorScheme } from "react-native";
 import { useTranslation } from "@localization";
 
 import styles from "./RankRowStyle";
-import { NO_VALUE, formatValue, unitBesideValue, unitLabel } from "../categoryLeaderboardFormat";
 import { Colors, withAlpha } from "@resources/GlobalStyling/colors";
 import { ThemedText, UserAvatar } from "@resources/ThemedComponents";
+import { NO_VALUE, formatValue, unitBesideValue, unitLabel } from "@utils/categoryFormat";
 
 /**
  * One place on the list from #4 down (from #1 on Progress). The list is the